SOURCE CODE/Startup-prediction/requirements.txt
flask
flask-cors
joblib
pandas
numpy
scikit-learn
imbalanced-learn
shap
vaderSentiment
requests
schedule
python-dotenv
matplotlib
seaborn
